Refactor TextToMeasurements to separate component
The logic for handling text conversion to measurements was extracted from the main page file into a dedicated component. This improves code organization and reusability, allowing for cleaner separation of logic and presentation layers within the application. Additionally, the main page now employs this new component for rendering the conversion functionality.

Took 1 minute
